feature: rework group printing
