+++ pkg/group/group_test.go
@@ -407,85 +407,3 @@ func Test_Reload(t *testing.T) {
 		})
 	}
 }
-
-func Test_Print(t *testing.T) {
-	type args struct {
-		State *state.State
-		Group []config.GroupItem
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "should print without an error",
-			args: args{
-				State: &state.State{
-					Group: state.Group{
-						Active: "dev",
-					},
-				},
-				Group: []config.GroupItem{
-					{
-						Name: "dev",
-						Context: config.Context{
-							Default: "kind-dev",
-						},
-						Sources: []string{
-							"dev",
-						},
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "should print without an error, even when there is no group",
-			args: args{
-				State: &state.State{},
-				Group: []config.GroupItem{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "should print without an error, even when there is no active group",
-			args: args{
-				State: &state.State{},
-				Group: []config.GroupItem{
-					{
-						Name: "dev",
-						Sources: []string{
-							"dev",
-						},
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "should print without an error, even when there is no source",
-			args: args{
-				State: &state.State{},
-				Group: []config.GroupItem{
-					{
-						Name: "dev",
-					},
-				},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client := Client{
-				State: tt.args.State,
-			}
-
-			err := client.Print(tt.args.Group...)
-			if !tt.wantErr && err != nil {
-				t.Errorf("err: '%v'", err)
-			}
-		})
-	}
-}
